Move net_http_persistent integration tests to the appropriate file
